TrabalhoFinal/RaphaelNagaoFerreira/pyproject.toml
[project]
name = "covidcxr4classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kagglehub>=0.3.13",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "scikit-learn>=1.7.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "typer>=0.20.0",
]
